pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "triviador"
version = "1.0.0"
description = "simple triviador using pygame"
requires-python = ">=3.10"
dependencies = [
    "pygame-ce>=2.5.0",
]

[project.scripts]
triviador = "triviador.app:main"

[tool.setuptools.packages.find]
include = ["triviador*"]

[tool.setuptools.package-data]
"triviador.data" = ["*.json", "*.txt"]

[tool.pytest.ini_options]
testpaths = ["tests"]
